requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
requests>=2.31.0
lxml>=4.9.0
google-generativeai>=0.3.0
pdf2image>=1.16.0
Pillow>=10.0.0
python-docx>=1.1.0

